pyproject.toml
[project]
name = "domain-checker"
version = "0.0.0"
dependencies = [
  "tqdm",
  "aiodns",
  "dnspython",
  "asyncio",
]
